Triplicate MS Word Macros virus
Triplicate was an early virus that took advantage of an oversight by Microsoft when they added the feature Macros to Excel, Word and PPT. It was developed by a University Student in Australia. Macros allowed the running of various methods and small programs made using Visual Basic.  I extracted this from a corrupted floppy disk that held a capstone project.  The word doc that contained this virus was created in the Jesse W. Mason Computer Lab at Georgia Tech in 1999. This and other evidence places the virus as an active problem for the Jesse W. Mason Computer Lab. Evidence will not be publically posted as it contains PII. 
-It is quite likely that it spread to other labs and machines). I can date that it was an active issue during the month of May in 1999